+++ yeat/cli.py
@@ -46,6 +46,7 @@ def run(
     dryrun="dry",
     downsample=0,
     coverage=150,
+    genomesize=0,
 ):
     snakefile = resource_filename("yeat", "Snakefile")
     r1 = Path(fastq1).resolve()
@@ -67,6 +68,7 @@ def run(
         dryrun=dryrun,
         downsample=downsample,
         coverage=coverage,
+        genomesize=genomesize,
     )
     success = snakemake(
         snakefile,
@@ -139,6 +141,14 @@ def get_parser(exit_on_error=True):
         help="target an average depth of coverage Cx when auto-downsampling; by default, C=150",
     )
     parser.add_argument(
+        "-g",
+        "--genome-size",
+        type=int,
+        metavar="G",
+        default=0,
+        help="genome size; by default, G=0",
+    )
+    parser.add_argument(
         "--init",
         action=InitAction,
         nargs=0,
@@ -162,4 +172,5 @@ def main(args=None):
         dryrun=args.dry_run,
         downsample=args.downsample,
         coverage=args.coverage,
+        genomesize=args.genome_size,
     )